# Dev env: Use ScyllaDB for faster testing