--- a/changelog.d/5-internal/scylladb-local
+++ b/changelog.d/5-internal/scylladb-local
@@ -1,0 +1,1 @@
+Use ScyllaDB in development environment in-place of Cassandra