… cassandra-schema.cql
… deploy/dockerephemeral/scylla.yaml