This PR updates the development environment: cassandra is replaced by [scylladb](https://www.scylladb.com/). On my machine™ I've seen substantinal performance improvements:

- Schema migrations are applied much faster
- Integration tests run faster: brig-integration runs in 60s (sylladb) vs 90s (cassandra)
- Less memory seems to be used

All integration tests suites pass locally for me.

## Checklist

 - [x] Add a new entry in an appropriate subdirectory of `changelog.d`
 - [x] Read and follow the [PR guidelines](https://docs.wire.com/developer/developer/pr-guidelines.html)