@pcapriotti 2023-11-02
Nice! Here are times I observed for setup and new integration:

#### old

```
make db-migrate  2.83s user 1.92s system 1% cpu 4:47.96 total
./dist/integration  52.89s user 22.20s system 51% cpu 2:26.75 total
```

#### scylla

```
make db-migrate  1.94s user 1.20s system 4% cpu 1:07.59 total
./dist/integration  54.83s user 22.66s system 48% cpu 2:39.92 total
```

So setup is more than 4x faster. However, the new integration tests are about the same. I guess that's because the runtime is dominated by spawning new backends and waiting for services to come up. Overall, it seems like a win.

What worries me is that I see a lot of:
```
[nginz@d1.example.com] 2023/11/02 14:41:41 [emerg] 130444#130444: io_setup() failed (11: Resource temporarily unavailable)
```
with this patch. Any idea what might be causing this?