@elland 2023-11-06
For me it had almost no effect. After running it using `time` and `hyperfine`, I've not found a significant performance difference between the two. I did however get a much faster `db-reset` result. Since it has no detrimental effect for my own env and clearly shows an improvement for others, sounds like a good idea to me. ✅