@smatting 2023-11-06
This is not good to merge yet. When running `./services/run-services` I get
```
[cannon@b.example.com] cannon: Network.Socket.bind: resource busy (Address already in use)
[cannon@example.com] cannon: Network.Socket.bind: resource busy (Address already in use)
[nginz@b.example.com] 2023/11/06 11:28:37 [emerg] 44976#44976: io_setup() failed (11: Resource temporarily unavailable)
[nginz@b.example.com] 2023/11/06 11:28:37 [emerg] 44977#44977: io_setup() failed (11: Resource temporarily unavailable)
[nginz@b.example.com] 2023/11/06 11:28:37 [emerg] 44978#44978: io_setup() failed (11: Resource temporarily unavailable)
[nginz@b.example.com] 2023/11/06 11:28:37 [emerg] 44979#44979: io_setup() failed (11: Resource temporarily unavailable)
```